change component structure
