+++ app/components/GameComponent/index.js
@@ -1,0 +1,10 @@
+//finish import and exports 
+
+
+import GameBlock from 'GameBlock'
+
+
+
+
+
+
