+++ app/components/InfoComponent/index.js
@@ -1,0 +1,1 @@
+//finish import and exports
