+++ app/components/index.js
@@ -1,0 +1,1 @@
+import * from 'GameBlock.js'
